Frontends/AI-UI/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
httpx>=0.23.0
jinja2>=3.1.2
python-multipart>=0.0.5
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.19.0
mistralai
opencv-python